requirements.txt
iniconfig==2.3.0
packaging==26.2
pluggy==1.6.0
Pygments==2.20.0
pytest==9.0.3
python-dateutil==2.9.0.post0
six==1.17.0
